Added source file for MatrixCalc.
